% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{add_permission}
\alias{add_permission}
\title{Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions}
\usage{
add_permission(TopicArn, Label, AWSAccountId, ActionName)
}
\arguments{
\item{TopicArn}{[required] The ARN of the topic whose access control policy you wish to modify.}

\item{Label}{[required] A unique identifier for the new policy statement.}

\item{AWSAccountId}{[required] The AWS account IDs of the users (principals) who will be given access to the specified actions. The users must have AWS accounts, but do not need to be signed up for this service.}

\item{ActionName}{[required] The action you want to allow for the specified principal(s).

Valid values: any Amazon SNS action name.}
}
\description{
Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions.
}
\section{Accepted Parameters}{
\preformatted{add_permission(
  TopicArn = "string",
  Label = "string",
  AWSAccountId = list(
    "string"
  ),
  ActionName = list(
    "string"
  )
)
}
}

