% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{create_topic}
\alias{create_topic}
\title{Creates a topic to which notifications can be published}
\usage{
create_topic(Name)
}
\arguments{
\item{Name}{The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.}
}
\description{
Creates a topic to which notifications can be published. Users can create at most 100,000 topics. For more information, see \href{http://aws.amazon.com/sns/}{http://aws.amazon.com/sns}. This action is idempotent, so if the requester already owns a topic with the specified name, that topic's ARN is returned without creating a new topic.
}
\examples{

}
