% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{delete_endpoint}
\alias{delete_endpoint}
\title{Deletes the endpoint for a device and mobile app from Amazon SNS}
\usage{
delete_endpoint(
  EndpointArn = "string"
)
}
\arguments{
\item{EndpointArn}{[required] EndpointArn of endpoint to delete.}
}
\description{
Deletes the endpoint for a device and mobile app from Amazon SNS. This action is idempotent. For more information, see \href{http://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\details{
When you delete an endpoint that is also subscribed to a topic, then you must also unsubscribe the endpoint from the topic.
}
\examples{

}
