% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{delete_platform_application}
\alias{delete_platform_application}
\title{Deletes a platform application object for one of the supported push notification services, such as APNS and GCM}
\usage{
delete_platform_application(PlatformApplicationArn)
}
\arguments{
\item{PlatformApplicationArn}{[required] PlatformApplicationArn of platform application object to delete.}
}
\description{
Deletes a platform application object for one of the supported push notification services, such as APNS and GCM. For more information, see \href{http://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Accepted Parameters}{
\preformatted{delete_platform_application(
  PlatformApplicationArn = "string"
)
}
}

