% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{delete_topic}
\alias{delete_topic}
\title{Deletes a topic and all its subscriptions}
\usage{
delete_topic(
  TopicArn = "string"
)
}
\arguments{
\item{TopicArn}{[required] The ARN of the topic you want to delete.}
}
\description{
Deletes a topic and all its subscriptions. Deleting a topic might prevent some messages previously sent to the topic from being delivered to subscribers. This action is idempotent, so deleting a topic that does not exist does not result in an error.
}
\examples{

}
