% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{get_platform_application_attributes}
\alias{get_platform_application_attributes}
\title{Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM}
\usage{
get_platform_application_attributes(PlatformApplicationArn)
}
\arguments{
\item{PlatformApplicationArn}{[required] PlatformApplicationArn for GetPlatformApplicationAttributesInput.}
}
\description{
Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see \href{http://docs.aws.amazon.com/sns/latest/dg/SNSMobilePush.html}{Using Amazon SNS Mobile Push Notifications}.
}
\section{Accepted Parameters}{
\preformatted{get_platform_application_attributes(
  PlatformApplicationArn = "string"
)
}
}

