% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{get_subscription_attributes}
\alias{get_subscription_attributes}
\title{Returns all of the properties of a subscription}
\usage{
get_subscription_attributes(SubscriptionArn)
}
\arguments{
\item{SubscriptionArn}{The ARN of the subscription whose properties you want to get.}
}
\description{
Returns all of the properties of a subscription.
}
\examples{

}
