% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{opt_in_phone_number}
\alias{opt_in_phone_number}
\title{Use this request to opt in a phone number that is opted out, which enables you to resume sending SMS messages to the number}
\usage{
opt_in_phone_number(phoneNumber)
}
\arguments{
\item{phoneNumber}{[required] The phone number to opt in.}
}
\description{
Use this request to opt in a phone number that is opted out, which enables you to resume sending SMS messages to the number.
}
\details{
You can opt in a phone number only once every 30 days.
}
\section{Accepted Parameters}{
\preformatted{opt_in_phone_number(
  phoneNumber = "string"
)
}
}

