% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_package.R
\docType{package}
\name{paws.sns-package}
\alias{paws.sns}
\alias{paws.sns-package}
\title{paws.sns: Amazon Simple Notification Service}
\description{
Amazon Simple Notification Service (Amazon SNS) is a web
service that enables you to build distributed web-enabled
applications. Applications can use Amazon SNS to easily push real-time
notification messages to interested subscribers over multiple delivery
protocols. For more information about this product see
[http://aws.amazon.com/sns](http://aws.amazon.com/sns/). For detailed
information about Amazon SNS features and their associated API calls,
see the [Amazon SNS Developer
Guide](http://docs.aws.amazon.com/sns/latest/dg/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
