% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{remove_permission}
\alias{remove_permission}
\title{Removes a statement from a topic's access control policy}
\usage{
remove_permission(TopicArn, Label)
}
\arguments{
\item{TopicArn}{[required] The ARN of the topic whose access control policy you wish to modify.}

\item{Label}{[required] The unique label of the statement you want to remove.}
}
\description{
Removes a statement from a topic's access control policy.
}
\section{Accepted Parameters}{
\preformatted{remove_permission(
  TopicArn = "string",
  Label = "string"
)
}
}

