% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{set_subscription_attributes}
\alias{set_subscription_attributes}
\title{Allows a subscription owner to set an attribute of the subscription to a new value}
\usage{
set_subscription_attributes(SubscriptionArn, AttributeName,
  AttributeValue = NULL)
}
\arguments{
\item{SubscriptionArn}{[required] The ARN of the subscription to modify.}

\item{AttributeName}{[required] The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable.

Valid values: \code{DeliveryPolicy} | \code{FilterPolicy} | \code{RawMessageDelivery}}

\item{AttributeValue}{The new value for the attribute in JSON format.}
}
\description{
Allows a subscription owner to set an attribute of the subscription to a new value.
}
\section{Accepted Parameters}{
\preformatted{set_subscription_attributes(
  SubscriptionArn = "string",
  AttributeName = "string",
  AttributeValue = "string"
)
}
}

