% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sns_operations.R
\name{set_topic_attributes}
\alias{set_topic_attributes}
\title{Allows a topic owner to set an attribute of the topic to a new value}
\usage{
set_topic_attributes(TopicArn, AttributeName, AttributeValue = NULL)
}
\arguments{
\item{TopicArn}{[required] The ARN of the topic to modify.}

\item{AttributeName}{[required] The name of the attribute you want to set. Only a subset of the topic's attributes are mutable.

Valid values: \code{Policy} | \code{DisplayName} | \code{DeliveryPolicy}}

\item{AttributeValue}{The new value for the attribute.}
}
\description{
Allows a topic owner to set an attribute of the topic to a new value.
}
\section{Accepted Parameters}{
\preformatted{set_topic_attributes(
  TopicArn = "string",
  AttributeName = "string",
  AttributeValue = "string"
)
}
}

